// src/types/sidebar.ts
export enum StudentNavItems {
    HOME = "Home",
    JOB_OPPORTUNITIES = "Job Opportunities",
    ABOUT_US = "About Us",
    CAREER_TIPS = "Career Tips",
    JOBS_BY_CITY = "Jobs by City",
    JOBS_BY_SKILL = "Jobs by Skill",
    JOBS_BY_DESTINATION = "Jobs by Destination",
    JOBS_BY_COMPANY = "Jobs by Company",
    OTHER_OPPORTUNITIES = "Other Opportunities",
}

export enum JobsByCity {
    DELHI = "Jobs in Delhi",
    MUMBAI = "Jobs in Mumbai",
    BANGALORE = "Jobs in Bangalore",
    CHENNAI = "Jobs in Chennai",
    PUNE = "Jobs in Pune",
    VIEW_ALL = "View All",
}

export enum JobsBySkill {
    SALES_MARKETING = "Sales & Marketing Jobs",
    ACCOUNTING = "Accounting Jobs",
    ENGINEERING = "Engineering Jobs",
    FINANCE = "Finance Jobs",
    NETWORKING = "Networking Jobs",
    VIEW_ALL = "View All",
}
export enum JobsByDesignation {
    SOFTWARE_ENGINEER = "Software Engineer Jobs",
    JAVA_PROGRAMMER = "Java Programmer Jobs",
    HR = "HR Jobs",
    PRODUCT_MANAGER = "Product Manager Jobs",
    ACCOUNTANT = "Accountant Jobs",
    VIEW_ALL = "View All",
}
export enum JobsByCompany {
    INFOSYS = "Infosys Jobs",
    GENPACT = "Genpact Jobs",
    BOSCH = "Bosch Jobs",
    SIEMENS = "Siemens Jobs",
    TCS = "TCS Jobs",
    VIEW_ALL = "View All",
}
export enum OtherJobOpportunities {
    WORK_FROM_HOME = "Work from Home Jobs",
    FREELANCING = "Freelancing Jobs",
    PART_TIME = "Part-Time Jobs",
    ONLINE = "Online Jobs",
    DATA_ENTRY = "Data Entry Jobs",
}